organising the searcher files
